fix: SECTION_IDS を非クライアントモジュールに分離（TOC 固定セクション ID 修正）
SECTION_IDS 定数と TocSection 型が "use client" モジュール（table-of-contents.tsx）に
定義されていたため、サーバーコンポーネント（page.tsx）からインポートすると RSC 境界で
undefined になり、href="#undefined" がレンダリングされていた。

- toc-config.ts を新設し SECTION_IDS / TocSection を分離（"use client" なし）
- table-of-contents.tsx から再エクスポートして後方互換を維持
- page.tsx のインポート元を toc-config.ts に変更
- NarrativeSection の narrativeContent パスに id="section-narrative" を追加
